Make quickhull3d reftest work with domino
There is quite a lot of ground to cover till this works.  The whole thing is
pretty hackish at the moment, but it appears to load the quickhull
implementation.  Since the result is more browser-like than it used to be,
we probably get some canvas transformation we didn't get before.