Proof of concept for the use of video in CindyGL
The images array may now contain DOM elements in addition to strings
specifying URLs.  This allows us to pass a video into the CindyJS instance,
and from there on to CindyGL.  There are some places where the distinction
between an image and a video is noticable, namely in the dimension
attributes, so the corresponding code needed some adjustments.