Sage code to compute maximal distance between conic and Bézier approximation
Trying to decide which approach to conic drawing is better, I've written
some code to take the paths generated for the test cases, and compute the
maximal distance between conic and Bézier path for each of them.  The points
where distance is locally maximal are those where the tangent to the conic
is parallel to the tangent to the Bézier curve.

There is also a function called drawconic3, taken from dc34a73 including the
outer code.  It has problems in a number of test cases like 781, where it
incorrectly draws a pair of lines.